created separate folder structure for RatingsAndReviews components
